add error handling to user use cases

+++ ca/src/ReqTrack/Domain.Core.UseCases/Users/DeleteUser/DeleteUserUseCase.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using ReqTrack.Domain.Core.Exceptions;
 using ReqTrack.Domain.Core.Repositories;
 using ReqTrack.Domain.Core.Security;
 using ReqTrack.Domain.Core.UseCases.Boundary;
 using ReqTrack.Domain.Core.UseCases.Boundary.Interfaces;
+using ReqTrack.Domain.Core.UseCases.Exceptions;
+using ReqTrack.Domain.Core.UseCases.Users.ChangeInformation;
 
 namespace ReqTrack.Domain.Core.UseCases.Users.DeleteUser
 {
@@ -22,16 +25,40 @@ namespace ReqTrack.Domain.Core.UseCases.Users.DeleteUser
         {
             try
             {
-                var result = _userRepository.DeleteUser(request.UserId);
+                if (!_userRepository.DeleteUser(request.UserId))
+                {
+                    throw new Exception("Couldn't delete user");
+                }
 
                 output.Response = new DeleteUserResponse(ExecutionStatus.Success)
                 {
                     Message = "User deleted successfully",
                 };
             }
+            catch (RequestValidationException e)
+            {
+                output.Response = new DeleteUserResponse(ExecutionStatus.Failure)
+                {
+                    UserId = request.UserId,
+                    Message = $"Invalid request: {e.Message}",
+                    ValidationErrors = e.ValidationErrors,
+                };
+            }
+            catch (EntityNotFoundException e)
+            {
+                output.Response = new DeleteUserResponse(ExecutionStatus.Failure)
+                {
+                    UserId = request.UserId,
+                    Message = $"User not found: {e.Message}",
+                };
+            }
             catch (Exception e)
             {
-                output.Response = new DeleteUserResponse(ExecutionStatus.TechnicalError);
+                output.Response = new DeleteUserResponse(ExecutionStatus.Failure)
+                {
+                    UserId = request.UserId,
+                    Message = $"Tehnical error happend: {e.Message}",
+                };
             }
         }
     }
